requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
filelock==3.20.2
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
packaging==25.0
pillow==12.0.0
pycparser==2.23
regex==2025.11.3
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
sounddevice==0.5.3
sympy==1.14.0
tiktoken==0.12.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2
